Add usePlayList to handle a playlist into a newUrl, answers and nextSong
