+++ src/App.js
@@ -3,119 +3,40 @@ import styled from 'styled-components'
 import useAudio from './hooks/useAudio'
 import PlayPage from './pages/PlayPage'
 import playlist from './data/playlist.json'
+import usePlayList from './hooks/usePlayList'
 
 function App() {
-  const [songList, setSongList] = useState(shuffle([...playlist]))
-  const [newUrl, setNewUrl] = useState(null)
-  const [songId, setSongId] = useState('1449299027')
-  const [rightAnswer, setRightAnswer] = useState()
-  const [wrongAnswers, setWrongAnswers] = useState([
-    { title: 'Angel Eyes' },
-    { title: 'Backstreets' },
-  ])
-  const [answers, setAnswers] = useState([
-    { title: 'Button A', right: true },
-    { title: 'Button B', wrong: true },
-    { title: 'Button C', wrong: true },
-  ])
+  const [showAnswer, setShowAnswer] = useState(false)
+  const { newUrl, answers, nextSong } = usePlayList(playlist)
+  const { songUrl, toggle, stop, playing, duration } = useAudio()
+  const [loaded, setLoaded] = useState(false)
+  const [newAnswers, setNewAnswers] = useState(answers)
 
-  //check songId for valid song
   useEffect(() => {
-    const baseUrl = `https://itunes.apple.com/lookup?id=${songList[0].id}`
-    songList[0].id &&
-      fetch(baseUrl)
-        .then(res => res.json())
-        .then(data => data.results.length === 0 && removeSong(songList[0].id))
-        .catch(error => console.error(error))
-  })
-  // set next song
-  useEffect(() => {
-    songList.length > '0'
-      ? setSongList(removeSong(songId))
-      : setSongList(shuffle([...playlist]))
-  }, [songId])
-
-  //get the wrong answers
-  useEffect(() => {
-    if (rightAnswer) {
-      const interpret = rightAnswer.interpret
-      const baseUrl = `https://itunes.apple.com/search?term=${interpret}&entity=song`
-      fetch(baseUrl)
-        .then(res => res.json())
-        .then(data => {
-          let answer1
-          let answer2
-          do {
-            answer1 = {
-              title:
-                data.results[Math.floor(Math.random() * data.results.length)]
-                  .trackName,
-            }
-            answer2 = {
-              title:
-                data.results[Math.floor(Math.random() * data.results.length)]
-                  .trackName,
-            }
-          } while (
-            answer1.title === answer2.title ||
-            answer1.title === rightAnswer.title ||
-            answer2.title === rightAnswer.title ||
-            answer1.title === rightAnswer.interpret ||
-            answer2.title === rightAnswer.interpret
-          )
-          setWrongAnswers([answer1, answer2])
-        })
-        .catch(error => console.error(error))
-    }
-  }, [rightAnswer])
-
-  //Get right song and artistName
-  useEffect(() => {
-    const baseUrl = 'https://itunes.apple.com/lookup?id='
-    fetch(baseUrl + songId)
-      .then(res => res.json())
-      .then(data => {
-        setNewUrl(data.results[0].previewUrl)
-        setRightAnswer({
-          title: data.results[0].trackName,
-          right: true,
-          interpret: data.results[0].artistName,
-        })
-      })
-      .catch(error => console.error(error))
-  }, [songId])
+    answers && setLoaded(true)
+  }, [answers])
 
   //set a new song
   useEffect(() => {
-    songUrl(newUrl)
+    !playing && songUrl(newUrl)
   }, [newUrl])
 
-  const [showAnswer, setShowAnswer] = useState(false)
-
-  // set the answer object in shuffled order for the buttons
   useEffect(() => {
-    if (rightAnswer && !showAnswer) {
-      const answer = [
-        { title: wrongAnswers[0].title, wrong: true },
-        { title: rightAnswer.title, right: true },
-        { title: wrongAnswers[1].title, wrong: true },
-      ]
-      setAnswers(shuffle(answer))
-    }
-  }, [rightAnswer, wrongAnswers, showAnswer])
-
-  const { songUrl, toggle, stop, playing, duration } = useAudio()
+    setNewAnswers(answers)
+  }, [playing])
 
   return (
     <Container>
-      <PlayPage
-        showAnswer={showAnswer}
-        answers={answers}
-        onPlay={handlePlay}
-        onAnswer={handleAnswer}
-        playing={playing}
-        duration={duration}
-      />
+      {loaded && (
+        <PlayPage
+          showAnswer={showAnswer}
+          answers={newAnswers ? newAnswers : answers}
+          onPlay={handlePlay}
+          onAnswer={handleAnswer}
+          playing={playing}
+          duration={duration}
+        />
+      )}
     </Container>
   )
 
@@ -123,7 +44,7 @@ function App() {
     if (playing) {
       stop()
       setShowAnswer(true)
-      setSongId(songList[0].id)
+      nextSong()
     }
 
     return
@@ -131,25 +52,9 @@ function App() {
 
   function handlePlay() {
     !playing && toggle()
-    if (showAnswer) {
+    if (showAnswer && !playing) {
       setShowAnswer(false)
     }
-  }
-
-  //Fisher–Yates shuffle modern algorithm
-  function shuffle(array) {
-    for (let i = array.length - 1; i > 0; i--) {
-      const j = Math.floor(Math.random() * (i + 1))
-      ;[array[i], array[j]] = [array[j], array[i]]
-    }
-    return array
-  }
-
-  function removeSong(removeId) {
-    const index = songList.findIndex(({ id }) => id === removeId)
-    const newSongList = songList
-    const removedSong = newSongList.splice(index, 1)
-    return newSongList
   }
 }
